Overhaul

+++ BaseCampApi/Authorization.cs
@@ -9,7 +9,7 @@ namespace BaseCampApi {
 	public class Token : ApiEntry {
 		public string access_token;
 		public string refresh_token;
-		public int expires_in;
+		public long expires_in;
 	}
 
 	/// <summary>
